Added folder size functionality.
This is not working quite right, it is showing slightly less than the windows estimation. Needs checking.

+++ Source/LibraryFolder.h
@@ -20,7 +20,7 @@ class LibraryFolder
     
 protected:
     fs::directory_entry RootDirectory;
-    int FolderSize = 0;
+    double FolderSize = 0;
 
     void CalculateFolderSize();
